refactor(overlay): move each layer into it's specific folder
