+++ src/components/Overlay/VignetteLayer/VignetteLayer.test.tsx
